Moved the askJobCode method to the cron command base class.
